- DecimalRow for Item Price - DatePicker for auction end date - Int -> Double for item price - custom auction cell font change - DateConversion class for fast converstion between date in ISO String and Date

+++ AMS/AMS/MyAuctions/Remote Storage/AuctionDocument.swift
@@ -101,9 +101,9 @@ class AuctionDocument : Document {
             case "shippingDetails":
                 auction.shippingDetails = data.value as! String
             case "startingPrice":
-                auction.startingPrice = data.value as! Int
+                auction.startingPrice = data.value as! Double
             case "price":
-                auction.price = data.value as! Int
+                auction.price = data.value as! Double
             case "sellerId":
                 auction.sellerId = data.value as! String
             case "buyerId":
